AssemblyAI/requirements.txt
requests==2.31.0
pydub==0.25.1
ffmpeg-python==0.2.0 